backend-files/requirements.txt
uvicorn[standard]==0.23.2
tqdm==4.66.1
python-multipart==0.0.6
ffmpeg-python==0.2.0
fastapi==0.104.1
llvmlite==0.41.1
numba==0.58.0
openai-whisper==20231117
faster-whisper==0.10.0
torch==1.13.0+cpu
--extra-index-url https://download.pytorch.org/whl/cpu
